Expose `operator<<` overloads for the `CurrentDomain` and `NDRectangle` C and CPP APIs. (#5538)
This PR exposes the `operator<<` overloads for the `CurrentDomain` and
`NDRectangle` C and C++ APIs. This change will simplify and unify the
implementation of corresponding dump functions in the high-level APIs,
as everything is now consolidated in the Core.

---
TYPE: C_API
DESC: Introduce `tiledb_current_domain_dump_str` and
`tiledb_ndrectangle_dump_str`.

TYPE: CPP_API
DESC: Introduce `operator<<` overloads for the `CurrentDomain` and
`NDRectangle`.