chore: Extend array rapidcheck templates to support nullable and variable-length fields (#5511)
Resolves
[CORE-23](https://linear.app/tiledb/issue/CORE-23/enable-nullable-and-variable-length-attributes-in-tiledb-core).

#5417 added a suite of tests which create an array with random data, and
then read back the data and compare against the input.

Those tests were limited to single-valued non-nullable attributes. This
pull request extends their capabilities to support nullable attributes
and variable-length attributes (and nullable variable-length
attributes).

Previously the code all used `std::vector<T>` to represent the
single-valued attribute data. Now to represent the set of buffers needed
to write or read a user type, we use `struct query_buffers<T>` where `T`
some logical representation of the user type:
- some fundamental type `T` for a non-nullable single-valued cell
- `std::vector<T>` for a fundamental type `T` for a non-nullable
variable-length cell
- `std::optional<T>` for a fundamental type `T` for a nullable
single-valued cell
- `std::optional<std::vector<T>>` for a fundamental type `T` for a
nullable variable-length cell

The specializations of `query_buffers` contain the columnar buffers for
each of these.

The `run_execute` function which drives the queries does multi-part
reads into a large `query_buffers` which it then compares against the
expected result. To handle the variable-length data it is necessary to
update the `uint64_t outcursor` which previously just held the number of
cells already read to a tuple which contains the variable-length
position as well as the cell position.

Because this is done with static dispatch there's a lot of `std::apply`
which basically means "turn this tuple into a variadic function argument
list".

# Testing

One of the existing tests is updated to use a non-nullable
variable-length attribute and one is updated to use a nullable
variable-length attribute.

---
TYPE: NO_HISTORY
DESC: Enable nullable/variable data in array rapidcheck tests

---------

Co-authored-by: Ypatia Tsavliri <ypatia@tiledb.com>