URL encode workspace and teamspace for REST URIs. (#5544)
Follow up for #5499 to add URL encoding for the workspace and teamspace
components for REST URIs if we are talking to a 3.0 REST server. If we
are using legacy REST the namespace component is returned as-is since it
is not allowed to contain characters that require encoding.

---
TYPE: BUG
DESC: URL encode workspace and teamspace for REST URIs.