chore: compile with `-D_GLIBCXX_ASSERTIONS` when building with `TILEDB_ASSERTIONS` configuration (#5531)
Resolves CORE-239.

In the recent past we've observed the occasional out-of-bounds access to
`std::vector` in our windows CI. In debug builds, windows includes this
sort of assertion.

While the instances of this have been benign, there's definitely value
in these debug checks which we are happy to opt into. However, Windows
is not a familiar platform for many developers; we would like something
more widely accessible.

While it likely is not a perfect match, the build option
`GLIBCXX_ASSERTIONS` offers similar functionality, adding assertions to
C++ standard library class methods to ensure correct usage.

This pull request adds that as a build option when the
`TILEDB_ASSERTIONS` feature is used, and fixes the code which fails as a
result (which we observed in nightly debug windows CI).

---
TYPE: NO_HISTORY
DESC: Compile with `-D_GLIBCXX_ASSERTIONS` when `TILEDB_ASSERTIONS` is
on