chore: test_assert expects correct return code from libc assertion failure test (#5550)
After #5548 we continue to observe failures in Windows nightlies, but
now because of the following:
```
  106: C:\Program Files\Microsoft Visual Studio\2022\Enterprise\VC\Tools\MSVC\14.43.34808\include\vector(1916) : Assertion failed: vector subscript out of range
  106: retval is c0000409 (0xc0000409) from C:/build/test/ci/Debug/try_libc_assert.exe
```
The `test_assert` binary is not expecting this exit code, which does not
match the typical assert exit codes.

This exit code means `STATUS_STACK_BUFFER_OVERRUN` which does seem
appropriate (if perhaps a little imprecise) given that this is an
out-of-bounds access to a `std::vector`.

This pull request updates the `test_assert` logic to allow this exit
code for windows.

---
TYPE: NO_HISTORY
DESC: allow `STATUS_STACK_BUFFER_OVERRUN` return code in `test_assert`