chore: build TileDB core with the capability of using Apache DataFusion to evaluate query conditions (#5546)
Resolves CORE-24.

This pull request extends TileDB core with the capability of using
Apache DataFusion to evaluate query conditions. While it is not
specifically our goal to deploy that capability in a production
environment, the changes from this pull request flesh out the
capabilities of our build system's Rust integration, as well as
implementing a large number of building blocks which we will need for
the follow-up project of using Apache DataFusion to evaluate arbitrary
text-based predicates with custom functions.

# Functionality

This pull request adds a new configuration parameter
`sm.query.condition_evaluator` with two allowed values:
1) `"ast"` which does nothing.
2) `"datafusion"` which instructs the `QueryCondition` to convert its
syntax tree into a `datafusion::logical_expr::Expr` syntax tree, and
then convert the tiles to `arrow::record_batch::RecordBatch` and use
`datafusion` to evaluate the `Expr` upon them.

We may remove this configuration parameter in the future. For now it is
useful to test the construction of `RecordBatch` and the `datafusion`
expression evaluation.

In this initial stage, we map all TileDB data types to the `arrow` data
type which represents the underlying physical data type. Single-valued
cells are converted to an integral type or float; multi-valued cells are
converted to `FixedSizeList`; and variable-length cells are converted to
`LargeList`. Future work may want or need to implement more
sophisticated mappings, but for now this is sufficient to implement the
operations supported by `QueryCondition`.

# Build System

#5528 introduced the ability to build with Rust components. This pull
request builds upon this to integrate Rust components in a non-trivial
way.

The build system design is centralized around the following note from
the [`cxx`](https://cxx.rs/build/other.html) crate developers:
> If you need to link multiple Rust subsystems, you will need to
generate a single staticlib perhaps using lots of extern crate
statements to include multiple Rust rlibs. Multiple Rust staticlib files
are likely to conflict.

We would like to be able to link Rust code into the TileDB core objects
as well as each of the standalone unit tests, without creating a single
monolithic entity. We also have multiple crates with `#[cxx::bridge]`
modules which we would want to use from one or more of those
`staticlib`s.

To abide these constraints, we have one Rust `staticlib` crate for each
of those top-level targets. These crates live in
`tiledb/oxidize/staticlibs`. These crates can define their own
`#[cxx::bridge]` for use in (and only in) the final build target they
correspond to; and they can export any number of other Rust crates with
`#[cxx::bridge]` modules.

To build one of the final build targets, it is necessary to link the
Rust `staticlib` as well as the `.cc` object files produced by expanding
the `#[cxx::bridge]` modules.

We use the open-source
[`Corrosion`](https://corrosion-rs.github.io/corrosion/) package to
populate `cmake` build targets for each of the Rust `staticlib`s, and
then we have our own `cmake` functions `cxxbridge` and `oxidze` which
populate targets for those `.cc` files.

`cmake` expresses only dependencies on the Rust `staticlib` and on the
files generated by `#[cxx::bridge]` module expansion. Cargo is
responsible for managing the build of the Rust staticlibs themselves.


![cmake-cargo-tree](https://github.com/user-attachments/assets/2a7cbc34-388b-4281-8346-6c01709e0dff)

# Testing

We have run all regression testing with `sm.query.condition_evaluator =
"datafusion"`, though we don't intend to merge with this configuration.

In addition there are tests added to `unit_query_condition` which
evaluate a query condition using the current AST implementation, and
compare results against evaluating the same query condition using
datafusion. There are a number of examples which are driven by both C++
and by Rust. And then we also use Rust `proptest` on top of this,
generating arbitrary schemata, query conditions and test data from our
`tiledb-rs` repository.

---
TYPE: NO_HISTORY
DESC: build TileDB core with the capability of using Apache DataFusion
to evaluate query conditions