Fix `Subarray::add_range` to validate variable-size consistency between range and dimension. (#5569)
Previously, when adding a range to a dimension in `Subarray::add_range`,
there was no validation to ensure that the variable-size property of the
range matched that of the target dimension. This could lead to bugs,
especially for variable-length dimensions, as was the case in
https://github.com/TileDB-Inc/TileDB/pull/5490, which was later
addressed in https://github.com/TileDB-Inc/TileDB/pull/5560.

This PR adds a validation check to ensure that the `var_size` of the
range matches the `var_size` of the corresponding dimension. If there is
a mismatch, a `SubarrayException` is thrown early.

Closes CORE-283

---
TYPE: IMPROVEMENT
DESC: Fix `Subarray::add_range` to validate var_size consistency between
range and dimension.