Update array create to pass the storage path in payload (#5565)
In the new Server the storage path should not be passed in the request
endpoint URL, so we need to specify it in the request payload. In Group
create case this was already implemented, but for arrays it was not.

This PR fixes it by introducing a new capnp message,
`array_create_request`, to be used with the new TileDB server only. This
new message will include, along with `array_schema`, a `storage_uri`
that will be indicating to the server where the physical storage of the
array will be.

---
TYPE: IMPROVEMENT
DESC: Update array create to pass the storage path in payload

---------

Co-authored-by: Shaun Reed <shaunrd0@gmail.com>