Fix empty array name for URIs with trailing slash. (#5591)
If a URI is used with a trailing slash `URI::last_path_part` returns an
empty string, which results in an empty array name when constructing the
request URL. This ensures an array name is found for this case, and
updates REST CI to use folders for storing assets to avoid collisions in
the DB.

REST-CI tests are passing locally using this branch against
TileDB-Server.

---
TYPE: BUG
DESC: Fix empty array name for URIs with trailing slash.