Fix `StatusException::extract_status` to support the status outliving the exception.
Should fix ASAN failures.
So, in `Status`, the origin is not copied, but in `StatusException` it is? This leads to lifetime complications when going from the latter to the former (which we need until `Status` gets completely eliminated from the codebase). The `Status`' origin is most of the time if not always a constant string, so this should be safe.