Correctly reset budget when initial data already loadad (#5599)
In the Server we support retrying query submit with a larger
`sm.mem.total_budget` when the initially provisioned budget is not
enough for loading tile offsets, which has been sometimes the case in
real world use cases in the past in heavy queries/large datasets. This
has been tested to work with the `DenseReader`, but not with Sparse
readers.

As a result there is a defect in Sparse that went unnoticed: When a
query is retried with increased budget, but the loading of its initial
data (E.g. rtrees) has fit in memory in a previous try, we aren't
(rightfully) re-loading the initial data, but proceeding directly to
loading the offsets that previously didn't fit. However, the logic for
resetting the memory budget of the reader to the new increased budget is
within the `load_initial_data` method that is now skipped, so the new
budget was never applied in practice, and the query was stuck getting
retried forever from the server.

This PR is fixing the issue by making sure we always reset the budget to
the caller provided value.

Closes CX-83

---
TYPE: BUG
DESC: Correctly reset bugdet when initial data already loadad