Mark object libraries with `EXCLUDE_FROM_ALL`. (#5606)
~~Not all~~ Almost none of the object libraries are actually being used
by the shipping binaries, and building the Core with just a `make`
command will cause an excessive amount of files to be compiled.

I noticed this while building the vcpkg port, and added a reduced
version of this PR as a patch, causing the number of source files
compiled to almost halve.

The `TILEDB_CORE_OBJECTS` object library was not excluded from `all`,
since it gets used by the shipping binaries either way.

---
TYPE: BUILD
DESC: Improved build times when not specifying a build target.