Do not retry SSL-related Curl error codes. (#5618)
This PR updates `Curl::should_retry_based_on_curl_code` to not retry
most error codes related to SSL. These are usually the result of
misconfiguration in either end of the connection, and retrying them is
not likely to change anything.

---
TYPE: IMPROVEMENT
DESC: Do not retry most SSL-related failures in the REST client.