Always check return codes in `Curl::get_data` and throw (#5632)
We are missing checks for all return values inside `Curl::get_data` so
that we throw accordingly . There's a chance that such errors that go
unnoticed in
[get_capabilities_from_rest](https://github.com/TileDB-Inc/TileDB/blob/7e2b6b8eefe45afc93b70dd21ce71dd8951f1e05/tiledb/sm/rest/rest_client_remote.cc#L1669),
produce the errors we have been seeing from
`rest_capabilities_deserialize` because they produce empty `data`.

Because of this any connectivity error
[[ref](https://curl.se/libcurl/c/curl_easy_setopt.html)] to the REST
server would mask itself as a `RestCapabilitiesSerializationException`,
as `get_capabilities_from_rest` is always the first request sent to the
REST server, so the first time we try to actually attempt a network
transfer.

---
TYPE: IMPROVEMENT
DESC: Always check return codes in `Curl::get_data` and throw