Bump CI to Ubuntu 24.04 with GCC/G++ 13. (#5574)
This PR updates the CI Ubuntu images to version 24.04 and also upgrades
g++ and gcc to version 13.

Sadly, I couldn't find any workaround for the timing tests when ASAN is
enabled and libfaketime is present. I was able to reproduce the issue
locally on an Ubuntu machine using the same build flags as ASAN CI.
I think it's a good idea to skip them for this particular combination
for now.

Closes CORE-387

---
TYPE: BUILD
DESC: Bump ubuntu version to 24.04.

---------

Co-authored-by: Theodore Tsirpanis <teo@tsirpanis.gr>