Replace deprecated `macos-13` GHA image with latest x64 image (#5649)
`macos-13` will soon be removed from github actions
[[src](https://github.blog/changelog/2025-09-19-github-actions-macos-13-runner-image-is-closing-down/)].
We are replacing it in our nightlies and release workflows with
`macos-15-intel` which according to
https://github.com/actions/runner-images/issues/13045 is now available
for building on x64.

However, let's keep in mind that "Apple has discontinued support for the
x86_64 (Intel) architecture. GitHub will no longer support this
architecture on macOS after the macOS 15 runner image is retired in Fall
2027."
[[src](https://github.blog/changelog/2025-09-19-github-actions-macos-13-runner-image-is-closing-down/#notice-of-macos-x86_64-intel-architecture-deprecation)]

---
TYPE: NO_HISTORY

---------

Co-authored-by: Theodore Tsirpanis <teo@tsirpanis.gr>