Check size of non empty domain before access (#5650)
During schema deserialization, when deserializing the `non_empty_domain`
the code depends on the existence of the `sizes` list to decide whether
or not the range is of variable length and access the array without
checking its size.
To make the deserialization more robust and because the serialization
process is not controlled by core (for example an already serialized
schema may be received from a client during a REST call), the size of
the `sizes` list should also be checked before accessing the list.

---
TYPE: BUG 
DESC: Check `non_empty_domain.sizes` size before access