Fix deserialization of relative group member URI (#5654)
When testing relative members with the Server, we noticed the URIs
returned from certain group member getter APIs were incorrect. This is
because we used a constructor for `URI` that implicitly converts paths
to absolute paths.

This PR is fixing this issue by checking if the URI is relative and
skips absolute path conversion in that case. It also adds a regression
test.

---
TYPE: BUG
DESC: Fix deserialization of relative group member URI