chore: add `IndexedList::splice` method and other tweaks (#5670)
This pull request is made in support of #5655 which remains in draft;
the intent is to pull these pieces out of that in order to have a more
focused code review of both pull requests.

We have a container class `IndexedList` which is used to store
non-movable data (and thus not usable in `std::vector`) with indexing
(and thus not usable in `std::list`). The implementation uses both
`std::list` and `std::vector` underneath.

This pull request makes some improvements to this class:
- we add the `splice` method whose semantics are the same as those of
the `std::list` `splice` method. This method transfers elements from one
`IndexedList` to another. In support of this we add new `iterator` and
`const_iterator` types.
- we provide the constructor definition in the header so that
implementations do not have to declare their own specialization.
- we update the `resize` method to accept variadic copyable arguments
which will be used to initialize each of the new elements of the
container.

We also add unit tests for `splice` correctness, and `resize` avoiding
the move constructor.

---
TYPE: NO_HISTORY
DESC: Add `IndexedList::splice`