Update shuffle filters to use blosc2 and remove the vendored sources. (#5620)
This PR updates the byteshuffle and bitshuffle filters to use the
[C-blosc2](https://www.blosc.org/c-blosc2/c-blosc2.html) library for
their implementation. This lets us remove several files from the
`external` directory, and simplify the build system.

### TODO

* [x] Write a one-off test that validates that the bitshuffle
implementations of blosc2 and `kiyo-masui/bitshuffle` behave identically
for various input sizes. From a quick read they should, but we need to
be extra sure, in case the backwards compatibility tests do not catch a
difference. This test doesn't have to permanently live in source
control, and we can remove it before merging the PR.

---
TYPE: IMPROVEMENT
DESC: The shuffle and bitshuffle filter implementations are updated to
use blosc2.

---
TYPE: BUILD
DESC: Removed vendored sources of blosc and `kiyo-masui/bitshuffle`, and
added a dependency to blosc2.