chore: `passert` failure should flush logs (#5682)
Resolves CORE-325.

One of the concerns reported in
[CORE-290](https://linear.app/tiledb/issue/CORE-290/consolidation-stdbad-alloc)
was that the process aborting did not flush logs and consequently did
not leave a lot of clues about what happened.

We cannot always intercept events which lead to process shutdown, but
one place we have managed to do so is assertion failures in core-owned
code. #5514 added `iassert` and `passert` routines, which throws an
exception or halts the process in the event of a failure respectively;
#5532 added support to run arbitrary callbacks in the event of a
`passert` failure.

This pull request leverages that capability to flush all logs in the
event of a `passert` failure so that if a customer sees a `passert`
failure we will know that the logs are up to date.

---
TYPE: NO_HISTORY
DESC: Flush logs if there is a `passert` failure