[Regression] Build the AWS SDK with hidden symbols (#5690)
PR #5223 fixed a bug where Arrow was causing a [double
initialization](https://github.com/apache/arrow/issues/40262) of the AWS
SDK, specifically on macOS. Recently PR #5616 removed this fix, causing
the nightly VCF macOS CI against TileDB `main` to
[fail](https://github.com/TileDB-Inc/TileDB-VCF/issues/855).

To address the issue, I forked this repo and cherry-picking the PR #5223
commit to head. I've confirmed both on a macOS M2 EC2 instance and in a
[VCF branch](https://github.com/TileDB-Inc/TileDB-VCF/pull/856) that
this resolves the issue.

TYPE: NO_HISTORY

Co-authored-by: Theodore Tsirpanis <theodore.tsirpanis@tiledb.com>