Do not provide storage URIs in TileDB-Server REST requests. (#5687)
Support for custom storage paths was removed in TileDB-Server, causing
REST CI failures since core sets the URI field for array / group
creation requests. This updates the requests to not set the URI field if
we are talking to a 3.0 REST server.

---
TYPE: BUG
DESC: Do not provide storage URIs in TileDB-Server REST requests.