chore: fix `test_ci_asserts` on windows nightlyRr/core 416 test assert nonzero exit code windows (#5696)
Resolves CORE-416.

The `test_ci_asserts` unit test has been causing windows nightly runs to
fail quite mysteriously as `ctest` would report all tests passing but
still return an exit code of `-1` to the runner, failing the job.

It turns out that when the `-V` (verbose) flag is used, `ctest` does
some processing of the stdout/stderr of each unit test. Even if the
tests succeed with an exit status of zero, certain patterns in the
stdout or stderr can cause `ctest` to report failure anyway.

One of those patterns seemingly is `internal error`, which is output by
our `passert` failure. So, when `test_ci_asserts` would spawn a
subprocess which intentionally fails asserts, `internal error` was
printed to stderr and flagged by `ctest`, causing it to report failure.

This pull request fixes that by capturing the output of the
assertion-failing subprocess instead so it doesn't get reported to
stdout.

We also fix the new `passert` callback registration tests, whose
failures we did not notice because they only failed in this nightly job.

---
TYPE: NO_HISTORY
DESC: Fix test_ci_asserts