Add C/C++ API for getting REST data protocol (#5697)
Adds a new API that can be used to tell which Server/data model
the REST client is configured to reach out to. This is needed as there
are differences in the data model between legacy and new TileDB Server,
especially when it comes to group operations, so a user needs to be able
to tell in an easy way to which server she is talking to.

---
TYPE: C_API
DESC: Add C API for getting REST data protocol

TYPE: CPP_API
DESC: Add C++ API for getting REST data protocol