Update group URI to Logical URI coming from the Server (#5702)
In the Server it is allowed to open a group by providing its URI in
AST_ID format. This causes problems for relative URIs, because Core has
some existing logic to concatenate the group name with its relative
members URI, in order to present nice and absolute member URIs to the
users instead of relative ones. However, concatenating AST_IDs with
member paths results in invalid asset URIs, as you can't append a path
to an ID.

To avoid this, we could either resort to just always return the member
URI as is, without concatenating it: this would make sense but would
break existing user scripts and expectations across backends. Or to ask
the Server to update the group URI to a logical one in Array open so
that concatenation always produces valid path URIs. This PR is aiming to
do the second.

part of CLOUD-2737

---
TYPE: BUG
DESC: Update group URI to Logical URI coming from the Server