Error if storage location is provided in array / group create request.  (#5711)
This updates #5687 to avoid swallowing the error returned by REST if the
custom storage location is provided in the request body for array /
group creation. After this change we throw the exception client side
before sending the request, if we are talking to a 3.0 REST server.

---
TYPE: BUG
DESC: Error if storage location is provided in array / group create
request.

---------

Co-authored-by: Ypatia Tsavliri <ypatia@tiledb.com>